Reorgnaize agent files
